BinaryTree: extract proof methods (#3889)
* binaryTree: extract proof methods to functions

* binaryTree: use functions in class methods

* binaryTree: test functions

* export functions from index.ts

* Remove lint-happy underscores

* binarytree: remove underscores from parameter names

* binaryTree: verify rootHash in verifyBinaryProof

* binaryTree: remove fromProof and verifyProof methods

---------
