+++ packages/binarytree/src/index.ts
@@ -2,4 +2,5 @@ export * from './binaryTree.js'
 export * from './constructors.js'
 export * from './db/index.js'
 export * from './node/index.js'
+export * from './proof.js'
 export * from './types.js'
